requirements.txt
# LangChain Core Dependencies
langchain==0.3.14
langchain-core==0.3.29
langchain-community==0.3.13
langchain-openai==0.2.14

# Vector Store
faiss-cpu==1.9.0.post1

# Local Embeddings (Free!)
sentence-transformers==3.3.1

# Utilities
python-dotenv==1.0.1
rich==13.9.4

# Additional Dependencies
numpy==1.26.4
tiktoken==0.8.0

